+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/TextEditingTarget.java
@@ -2652,7 +2652,15 @@ public class TextEditingTarget implements
             ? prefs_.autoAppendNewline().getValue()
             : projConfig_.ensureTrailingNewline();
             
-      if (autoAppendNewline || fileType_.isPython())
+      // auto-append newlines for commonly-used R startup files
+      String path = StringUtil.notNull(docUpdateSentinel_.getPath());
+      boolean isStartupFile =
+            path.endsWith("/.Rprofile") ||
+            path.endsWith("/.Rprofile.site") ||
+            path.endsWith("/.Renviron") ||
+            path.endsWith("/.Renviron.site");
+      
+      if (autoAppendNewline || isStartupFile || fileType_.isPython())
       {
          String lastLine = docDisplay_.getLine(lineCount - 1);
          if (lastLine.length() != 0)